Progress on the new figures.  S0lved a few look and feel problems.
Since the move to markdown, we cannot control the size of the figures in PDF.
There seems to be no good way to do this - so I invented one.  I augment
the image lines with LaTeX figure size information:

![Personal Digital Assistant](height=1.0in@../images/pda)

Before the @ gors straight into the includegraphics line if we
are going to LaTeX

\includegraphics[width=0.25cm,height=0.25cm]{../images/pda.eps}

If we are going to HTML or EPUB this extra stuff needs to be
removed so I made a pre-html.py patcher.