Update peer2.json
Feedback from student. The default gline is by-month already yyyy-mm. 
https://www.coursera.org/learn/python-data-visualization/discussions/weeks/6/threads/KaNNURjCEeipWQqtKjEwpg

Also spotted a typo, "screen shot to the" replace "to" by "of".

I know this was mentioned before, and we were expecting a graph from you, for the by-year example. Not that you would write the gyear.py and put it in the zip. So I don't know if you want to remove gyear.py and out of the zip. Otherwise optional challenge is too easy,  for those that find it.