Update spider.py
Remove redundant line in code, if entry deleted in line above there is no record to update. error = 0 is never written to db. It's just causing students problems. Not accepting it was a feature not used, it doesn't do anything, it doeesn't matter if it's in the code block, it does run but has no effect.