added context=ctx parameter for use with https
The ctx that accounts for handling https sites had been defined in previous lines but not used. 
This change is in line with the parameters passed to urllib.request.urlopen() that account for both http and https sites, as defined in the lecture videos and other sample files provided by Dr. Chuck.