Clarify Exercise 8.4
My wife is currently doing the Python for Everyone specialization and
found Exercise 8.4 confusing for two reasons:

1. The assignment jumps directly into what code to write, without
clarifying what is the overall objective.
2. The assignment uses the word "list" with double meaning, once
referring to the list of words obtained after `split()`-ing a line, and
once referring to the list of unique words to output.

This patch improves the two points above.