Update re12.py
Student post, https://www.coursera.org/learn/python-network-data/discussions/all/threads/f5FO4dW4Eeun7grBzPuYww
The Details line is an integer, the . can be removed and the result doesn't change. Tested edit using mbox-short, mbox, check in editor find, count lines and it matches.

Update text as it is not number > 0, but the len, a result is found.