Update urllinks.py with comments
SSL is deprecated, although the package is still named SSL and many companies still call their certs SSL certificates. It's useful to surface this so people learning about security are aware of the shift, but comfortable using the SSL package. 

Also Python developers should become aware of the normal install mechanism, which takes advantage of the latest version. This encourages better security practices.